+++ src/print_node.h
@@ -22,9 +22,10 @@
 #include "../inst/include/rxode2parse.h"
 #include "../inst/include/rxode2parseSbuf.h"
 #include "tran.h"
+#include "strncmpi.h"
 
 static inline int nodeTime(char *value) {
-  if (!strcmp("time",value)){
+  if (!rxstrcmpi("time",value)){
     aAppendN("t", 1);
     sAppendN(&sbt, "t", 1);
     return 1;